Fixed db error for makepost
poster -> posterID
